style(src): resolve redundant qualifiers, casts, inline, boolean args, elaborated types, lambda params
Remove redundant namespace qualifiers inside projectv::render/voxel,
unnecessary static_casts, inline specifiers, boolean expression arguments,
elaborated 'struct' keywords, and empty lambda parameter lists. Keep
explicit narrowing casts and ambiguous-call qualifiers with noinspection
comments where CLion flags are false positives.

Build green; 43/43 ctest pass.

+++ src/render/Cloudscape.cpp
@@ -90,8 +90,8 @@ float Fbm2D(float x, float y, int octaves)
 
 std::vector<uint8_t> GenerateCloudscapeNoiseR8()
 {
-	std::vector<uint8_t> data(static_cast<std::size_t>(projectv::render::kCloudscapeNoiseTextureSize *
-							  projectv::render::kCloudscapeNoiseTextureSize));
+	std::vector<uint8_t> data(projectv::render::kCloudscapeNoiseTextureSize *
+							  projectv::render::kCloudscapeNoiseTextureSize);
 	for (uint32_t y = 0; y < projectv::render::kCloudscapeNoiseTextureSize; ++y) {
 		for (uint32_t x = 0; x < projectv::render::kCloudscapeNoiseTextureSize; ++x) {
 			const float u = static_cast<float>(x) / static_cast<float>(projectv::render::kCloudscapeNoiseTextureSize);
